+++ packages/bundle/src/index-es5.ts
@@ -30,6 +30,7 @@ import 'whatwg-fetch';
 import { version } from './index-minimal';
 import addVersion from './addVersion';
 import defaultCreateDirectLine from './createDirectLine';
+import defaultCreateDirectLineAppServiceExtension from './createDirectLineAppServiceExtension';
 
 export * from './index';
 
@@ -38,12 +39,23 @@ export const createDirectLine = options => {
     console.warn(
       'Web Chat: Developers are not currently allowed to set botAgent in the createDirectLine function. See https://github.com/microsoft/BotFramework-WebChat/issues/2119 for more details.'
     );
+
   return defaultCreateDirectLine({ ...options, botAgent: `WebChat/${version} (ES5)` });
+};
+
+export const createDirectLineAppServiceExtension = options => {
+  options.botAgent &&
+    console.warn(
+      'Web Chat: Developers are not currently allowed to set botAgent in the createDirectLine function. See https://github.com/microsoft/BotFramework-WebChat/issues/2119 for more details.'
+    );
+
+  return defaultCreateDirectLineAppServiceExtension({ ...options, botAgent: `WebChat/${version} (ES5)` });
 };
 
 window['WebChat'] = {
   ...window['WebChat'],
-  createDirectLine
+  createDirectLine,
+  createDirectLineAppServiceExtension
 };
 
 addVersion('full-es5');