+++ packages/bundle/src/index-minimal.ts
@@ -15,6 +15,7 @@ import addVersion from './addVersion';
 import coreRenderWebChat from './renderWebChat';
 import createBrowserWebSpeechPonyfillFactory from './createBrowserWebSpeechPonyfillFactory';
 import defaultCreateDirectLine from './createDirectLine';
+import defaultCreateDirectLineAppServiceExtension from './createDirectLineAppServiceExtension';
 
 const renderWebChat = coreRenderWebChat.bind(null, ReactWebChat);
 
@@ -23,7 +24,17 @@ export const createDirectLine = options => {
     console.warn(
       'Web Chat: Developers are not currently allowed to set botAgent in the createDirectLine function. See https://github.com/microsoft/BotFramework-WebChat/issues/2119 for more details.'
     );
+
   return defaultCreateDirectLine({ ...options, botAgent: `WebChat/${version} (Minimal)` });
+};
+
+export const createDirectLineAppServiceExtension = options => {
+  options.botAgent &&
+    console.warn(
+      'Web Chat: Developers are not currently allowed to set botAgent in the createDirectLine function. See https://github.com/microsoft/BotFramework-WebChat/issues/2119 for more details.'
+    );
+
+  return defaultCreateDirectLineAppServiceExtension({ ...options, botAgent: `WebChat/${version} (Minimal)` });
 };
 
 export default ReactWebChat;
@@ -48,6 +59,7 @@ window['WebChat'] = {
   Constants,
   createBrowserWebSpeechPonyfillFactory,
   createDirectLine,
+  createDirectLineAppServiceExtension,
   createStore,
   createStyleSet,
   hooks,