+++ packages/testharness/src/token/fetchDirectLineAppServiceExtensionToken.js
@@ -1,0 +1,11 @@
+export default async function(url = 'https://webchat-mockbot2.azurewebsites.net/api/token/directlinease') {
+  const res = await fetch(url, { method: 'POST' });
+
+  if (!res.ok) {
+    throw new Error('Failed to fetch Direct Line App Service Extension token.');
+  }
+
+  const { token } = await res.json();
+
+  return token;
+}